requirements.txt
asgiref
black
Django
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2
psycopg2-binary
dj-database-url
python-decouple
django-cors-headers
whitenoise
django-filter
pytest
pytest-django
factory-boy
mock
Faker
django-cloudinary-storage
cloudinary
Pillow
drf-spectacular
channels
channels-redis
django-channels
daphne
redis
python-dotenv
google-generativeai
